Moving upload file to assesment

+++ meteor-app/imports/api/parts/methods.js
@@ -6,15 +6,15 @@ const debug = require('debug')('b2b:parts')
 
 function calcRetail(price){
   if (price <= 6000) {
-    const retailPrice = (parseInt(price, 10) * 100) * 2
+    const retailPrice = (parseFloat(price, 10) * 100) * 2
     return retailPrice
   }
   else if (price > 6000 && price <= 10000) {
-    const retailPrice = (parseInt(price, 10) * 100) * 1.5
+    const retailPrice = (parseFloat(price, 10) * 100) * 1.5
     return retailPrice
   }
   else if (price > 10000) {
-    const retailPrice = (parseInt(price, 10) * 100) * 1.3
+    const retailPrice = (parseFloat(price, 10) * 100) * 1.3
     return retailPrice
   }
   else {
